refactor root command
