+++ cmd/root/logo.go
@@ -1,0 +1,18 @@
+package root
+
+const logo = `
+                      ~~~        
+            ---       ~~~~~~~      
+     °      -------   ~~~~~~~~~~  
+     °°°°   ----------- ~~~~~~~~~
+     °°°°°°° ----------- ~~~~~~~~       _   _                                              
+     °°°°°°°°°°  ------- ~~~~~~~~      | \ | |                                         
+     °°°°°°°°°°°°° -----  ~~~~~~~      |  \| |_   _ _ __ ___   ___ _ __ ___  _   _ ___
+     °°°°°°°°°°°°° -----  ~~~~~~~      | . ` + "`" + ` | | | | '_ ` + "`" + ` _ \ / _ \ '__/ _ \| | | / __|
+     °°°°°°°°°°°°° -----     ~~~~      | |\  | |_| | | | | | |  __/ | | (_) | |_| \__ \
+     °°°°°°°°°°°°°  ----       ~~      |_| \_|\__,_|_| |_| |_|\___|_|  \___/ \__,_|___/
+        °°°°°°°°°°    --
+          °°°°°°°°    
+             °°°°°   
+                °°     
+`
